telegram_mock_bot/requirements.txt
Flask==2.3.2
python-telegram-bot==20.7
python-dotenv==1.0.1